@@ -9,15 +9,14 @@
     "@fortawesome/free-regular-svg-icons": "^5.15.3",
     "@fortawesome/free-solid-svg-icons": "^5.15.3",
     "@fortawesome/react-fontawesome": "^0.1.14",
-    "@reduxjs/toolkit": "^1.5.1",
+    "@reduxjs/toolkit": "^1.9.7",
     "@testing-library/jest-dom": "^5.12.0",
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
     "@types/jest": "^26.0.22",
     "@types/node": "^12.20.10",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^17.0.3",
-    "@types/react-redux": "^7.1.16",
+    "@types/react": "^18.3.5",
+    "@types/react-dom": "^18.3.0",
     "antlr4": "^4.9.2",
     "bootstrap": "^5.3.3",
     "convert-units": "^3.0.0-beta.2",
@@ -27,13 +26,13 @@
     "lodash": "^4.17.21",
     "minisearch": "^3.1.0",
     "numeral": "^2.0.6",
-    "react": "^17.0.2",
+    "react": "^18.3.1",
     "react-bootstrap": "^2.10.4",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
+    "react-dom": "^18.3.1",
+    "react-redux": "^8.1.3",
     "react-scripts": "5.0.1",
     "redux": "^4.1.0",
-    "typescript": "^4.2.4",
+    "typescript": "^4.9.4",
     "web-vitals": "^1.1.1"
   },
   "scripts": {
@@ -71,6 +70,7 @@
     "serve": "^14.2.3"
   },
   "overrides": {
-    "autoprefixer": "10.4.5"
+    "autoprefixer": "10.4.5",
+    "@types/react": "^18.3.5"
   }
 }
