@@ -1,21 +1,21 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
 import 'bootstrap/dist/css/bootstrap.css';
+import React from 'react';
 import './index.css';
 import reportWebVitals from './reportWebVitals';
 
+import { createRoot } from 'react-dom/client';
 import { Provider } from "react-redux";
-import store from "./app/store";
 import App from './App';
+import store from "./app/store";
 
-
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container!);
+root.render(
   <React.StrictMode>
     <Provider store={store}>
       <App />
     </Provider>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
